WhatsAppBot/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
twilio==8.9.0
python-dotenv==1.0.0
google-genai
requests==2.31.0
python-multipart==0.0.6